requirements.txt
Django==4.0
spotipy==2.19.0
python-dotenv==0.19.2
scikit-learn==1.0.1
pandas==1.3.5
numpy==1.21.5